refactors code according to solutions videos for re-sorting queue_items and for destroying a queue item

+++ app/controllers/queue_items_controller.rb
@@ -13,17 +13,30 @@
 
   def destroy
     queue_item = QueueItem.find(params[:id])
-    queue_item.destroy
+    queue_item.destroy if current_user.queue_items.include?(queue_item)
     redirect_to queue_items_path
   end
 
   def sort
-    queue_items = params[:queue_items]
-    #binding.pry
-    queue_items.each_with_index do |id, index|
-      QueueItem.where({position: index+1}, {id: id}).to_a
+    begin
+      ActiveRecord::Base.transaction do
+        params[:queue_items].each do |queue_item_data|
+          queue_item = QueueItem.find(queue_item_data["id"])
+          queue_item.update!(position: queue_item_data["position"]) if queue_item.user == current_user
+        end
+      end
+    rescue ActiveRecord::RecordInvalid
+      flash[:notice] = "Please only use whole numbers to update the queue" 
+      redirect_to queue_items_path
+      return
     end
-    redirect_to(queue_items_path)    
+
+    current_user.queue_items.each_with_index do |queue_item, index|
+      queue_item.update(position: index+1)
+    end
+
+    redirect_to(queue_items_path)
+    
   end
 
 end
